refactor(): clean up a bunch of strict tsc warnings
A number of DAO calls that would silently fail with undefined errors
when the server is not configured yet will now explicitly throw an
OnmsError.

Also note: A number of methods and properties that used to be assumed
to be required can now be undefined.  In the real world these properties
would already be optional, but they are now explicity specified in
the TypeScript signatures.